This is how I have right now the script on my PC, without my personal installations (i.e. banshee, emacs).
I have added a brief description of the script, and I plan to make it
ignore some installation scripts and execute others. That way, I'll be
able to add my own scripts and don't affect other people, unless it's
specified.

I've also added an empty script with development options on purpose. I
plan to implement it with just a few lines that enable the core dump
files generation, and some other stuff I have on the PC but still need
to organise neatly.