setup_linux.sh: development tools must be installed as a last step
It doesn't matter if this is performed before or after the system update,
but it can't be run when parsing the arguments.

Without this commit the command `./setup_linux.sh --devel --users foo`
would try to run `setup_linux.sh` first, which wouldn't even have the
users variable set.

Therefore, from now on, all the optional choices retrieved from the
arguments and run the appropriate ones only in the final steps.