all: Port to Puppet
It's easier to deploy configuration files and packages using Puppet. Even
by using it, we still need to make sure it's installed in system.

One shortcoming of this tool is that, by trying to provide a high
abstraction of what commands the system should use in order to deploy
packages and files, it doesn't let a user step in and execute system
commands and retrieve their output. That's why I've needed to add some
extra lines to substitute "on-the-fly" the home directory in all the
manifest files.