commandline/files/bashrc: Add wrapper for autossh
I've been using this simple helper for a while now because autossh can't
handle some network problems and I really needed something that would keep
retrying to keep the ssh tunnel open indefinitely.

This function is meant to be used as:
$ infinite_ssh 23000 user@server

I've assumed the local port is 22 because I never change it.

Note that this, as well as autossh, requires a password-less setup.