tmux.conf: Set up initial windows and panes
I find this useful because every time I open a new tmux session, I set up
the same windows all over again.

This commit adds the necessary lines to its config file so that I just need
to add my windows and commands.

In the initial windows, I've set the option "remain-on-exit" to "on" so
that the windows don't close as soon as I kill the process. I sometimes
need to restart a process so that, for example, it takes a new
configuration.

The only short-coming to this change is that now, running tmux with the
plain command "tmux" is equivalent to "tmux new-session". Therefore, in
order to run tmux and attach to it as I did before, now I need to run
"tmux attach".