devel: New aliases git-diff-prev and git-diff-branch
Sometimes I keep committing my changes and amending the commit as things
start working. Then, if I do git diff (to run meld), it doesn't show any
diff because all my work is already committed. That's why I need an alias
"git diff-prev", so that I don't need to figure out the reference of the
previous commit and do "git diff $ref".

"git-diff-branch", on the other hand, does something similar but it looks
for the newest commit in the intended branch. This is to see what changes
I've made in my whole branch. This is useful when a "master branch" has
been pulled and there are more changes in-between.