common_user_files: Ensure presence of common required files
~/.bashrc and ~/.bash_profile are required by some things in both modules
"commandline" and "devel". It would be much simpler to ensure the presence
of ~/.bashrc in both of them, but Puppet doesn't admit duplicate entries
because you could end up with contradictory information all over the place.

For this reason, I've added this new module, and other modules that need to
use these files will depend on it.