setup_linux.sh: Patch ~/.gitconfig with ~/.gitconfig.extra after execution
This is necessary because when the script sets up all the system, when it
sets up the file ~/.gitconfig, it replaces it with the file I have in the
module devel.

Another option would have been to apply this patch with the same logic to
append the three lines for the external diff directly onto ~/.gitconfig.
However, it's cleaner to do it this way because any change made on the
config file will be cleaned up and restored to the previous contents it
should have, which is the purpose of this script.

All a user needs to do is add the contents that wants to be persistent in
a file called ~/.gitconfig.extra