setup_linux.sh: Fix bad refactoring
In a previous commit, I added some logic to run patch_{before,after}_file()
functions. However, in the process I ended up having the function
patch_after_devel.pp() duplicated.

This commit moves the logic that used to be in the original function into
the new one and gets rid of the duplicated (old) one.