backup: Add code to create backup in external hard drive
By default, with no changes performed in the file, it will not attempt to
do this, but if the variable "CANDIDATE_PATH" is set to the path where the
external hard drive has been mounted, the backup will be done.

I've chosen to do it this way instead of taking an argument because it's
easier to just type the command line "backup", rather than remembering
whether the arguments is preceded by "-p" or something different.