commandline: tmux.conf doesn't capture the mouse
This was introduced in 6c0022f023cb90b010b7058fe168f712680b58b1 to make it
easy to occasionally share my PC with colleagues, but it turns out to be
messing it up. In particular, I can't select text to copy anymore, which
makes it very difficult to work with it.